Publicly accessible link to the file in DropBox
This file is a combination of the default password.lst file with John The Ripper (only about 3,500 passwords) with cain-and-able.txt (obtained here: https://github.com/danielmiessler/SecLists/blob/master/Passwords/Software/cain-and-abel.txt) and rockyou.txt (obtained here: https://github.com/brannondorsey/naive-hashcat/releases/download/data/rockyou.txt). I combined them, and then cleaned up the file and tested it by breaking a cryptic password for a PDF (selected a password from the list to confirm that it was working as expected).

This has resulted in a password list with just over 14.5 million passwords.